zson

+++ ft_memset.c
@@ -6,16 +6,16 @@
 /*   By: mutopal <mutopal@student.42istanbul.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/20 17:30:26 by mutopal           #+#    #+#             */
-/*   Updated: 2025/06/10 16:30:01 by mutopal          ###   ########.fr       */
+/*   Updated: 2025/06/26 22:37:25 by mutopal          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memset(void *ptr, int value, int num)
+void	*ft_memset(void *ptr, int value, size_t num)
 {
-	int				i;
-	unsigned char	*p;
+	size_t				i;
+	unsigned char		*p;
 
 	p = (unsigned char *)ptr;
 	i = 0;
