+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: mutopal <mutopal@student.42istanbul.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/19 23:58:14 by mutopal           #+#    #+#             */
-/*   Updated: 2025/06/20 22:24:46 by mutopal          ###   ########.fr       */
+/*   Updated: 2025/06/28 15:13:58 by mutopal          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
 	size_t	i;
